pyproject.toml
[tool.poetry]
name = "store-with-payment"
version = "0.1.0"
description = ""
authors = ["TimurMeshchenko <matsuokisho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
stripe = "^9.1.0"
pillow = "^10.3.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
